Integrate with asset live data
