+++ js_modules/dagster-ui/packages/ui-core/src/asset-data/AssetBaseDataQueries.tsx
@@ -1,0 +1,113 @@
+import {gql} from '../apollo-client';
+
+export const ASSET_LATEST_INFO_FRAGMENT = gql`
+  fragment AssetLatestInfoFragment on AssetLatestInfo {
+    id
+    assetKey {
+      path
+    }
+    unstartedRunIds
+    inProgressRunIds
+    latestRun {
+      id
+      ...AssetLatestInfoRun
+    }
+  }
+
+  fragment AssetLatestInfoRun on Run {
+    status
+    endTime
+    id
+  }
+`;
+
+export const ASSET_NODE_LIVE_FRAGMENT = gql`
+  fragment AssetNodeLiveFragment on AssetNode {
+    id
+    opNames
+    repository {
+      id
+    }
+    assetKey {
+      path
+    }
+    assetMaterializations(limit: 1) {
+      ...AssetNodeLiveMaterialization
+    }
+    assetObservations(limit: 1) {
+      ...AssetNodeLiveObservation
+    }
+    assetChecksOrError {
+      ... on AssetChecks {
+        checks {
+          ...AssetCheckLiveFragment
+        }
+      }
+    }
+    partitionStats {
+      numMaterialized
+      numMaterializing
+      numPartitions
+      numFailed
+    }
+  }
+
+  fragment AssetNodeLiveMaterialization on MaterializationEvent {
+    timestamp
+    runId
+  }
+
+  fragment AssetNodeLiveObservation on ObservationEvent {
+    timestamp
+    runId
+  }
+
+  fragment AssetCheckLiveFragment on AssetCheck {
+    name
+    canExecuteIndividually
+    executionForLatestMaterialization {
+      id
+      runId
+      status
+      timestamp
+      stepKey
+      evaluation {
+        severity
+      }
+    }
+  }
+`;
+
+export const ASSETS_GRAPH_LIVE_QUERY = gql`
+  query AssetGraphLiveQuery($assetKeys: [AssetKeyInput!]!) {
+    assetNodes(assetKeys: $assetKeys, loadMaterializations: true) {
+      id
+      ...AssetNodeLiveFragment
+    }
+    assetsLatestInfo(assetKeys: $assetKeys) {
+      id
+      ...AssetLatestInfoFragment
+    }
+  }
+
+  ${ASSET_NODE_LIVE_FRAGMENT}
+  ${ASSET_LATEST_INFO_FRAGMENT}
+`;
+
+export const ASSETS_FRESHNESS_INFO_QUERY = gql`
+  query AssetsFreshnessInfoQuery($assetKeys: [AssetKeyInput!]!) {
+    assetNodes(assetKeys: $assetKeys) {
+      id
+      assetKey {
+        path
+      }
+      freshnessInfo {
+        ...AssetNodeLiveFreshnessInfoFragment
+      }
+    }
+  }
+
+  fragment AssetNodeLiveFreshnessInfoFragment on AssetFreshnessInfo {
+    currentMinutesLate
+  }
+`;
