+++ js_modules/dagster-ui/packages/ui-core/src/asset-data/AssetLiveDataProvider.tsx
@@ -1,14 +1,14 @@
 import uniq from 'lodash/uniq';
 import React, {useCallback, useMemo, useRef} from 'react';
+import {LiveDataForNodeWithStaleData} from 'shared/asset-graph/LiveDataForNode.oss';
 
 import {AssetBaseData, __resetForJest as __resetBaseData} from './AssetBaseDataProvider';
-import {AssetSLAData} from './AssetSLADataProvider';
 import {
   AssetStaleStatusData,
   __resetForJest as __resetStaleData,
 } from './AssetStaleStatusDataProvider';
 import {observeAssetEventsInRuns} from '../asset-graph/AssetRunLogObserver';
-import {LiveDataForNodeWithStaleData, tokenForAssetKey} from '../asset-graph/Utils';
+import {tokenForAssetKey} from '../asset-graph/Utils';
 import {AssetKeyInput} from '../graphql/types';
 import {LiveDataPollRateContext} from '../live-data-provider/LiveDataProvider';
 import {LiveDataThreadID} from '../live-data-provider/LiveDataThread';
@@ -156,9 +156,7 @@ export const AssetLiveDataProvider = ({children}: {children: React.ReactNode}) =
 
   return (
     <AssetBaseData.LiveDataProvider>
-      <AssetStaleStatusData.LiveDataProvider>
-        <AssetSLAData.LiveDataProvider>{children}</AssetSLAData.LiveDataProvider>
-      </AssetStaleStatusData.LiveDataProvider>
+      <AssetStaleStatusData.LiveDataProvider>{children}</AssetStaleStatusData.LiveDataProvider>
     </AssetBaseData.LiveDataProvider>
   );
 };
