--- js_modules/dagster-ui/packages/ui-core/src/asset-data/getAssetSLAQueryResponse.oss.tsx
@@ -1,5 +1,0 @@
-import {ApolloClient} from '../apollo-client';
-
-export const getAssetSLAQueryResponse = (_assetKeys: string[], _client: ApolloClient<any>) => {
-  return {};
-};
