+++ js_modules/dagster-ui/packages/ui-core/src/asset-data/queryAndBuildAssetBaseData.oss.tsx
@@ -1,0 +1,11 @@
+import {LiveDataForNode} from 'shared/asset-graph/LiveDataForNode.oss';
+
+import {queryAndBuildAssetBaseOSSData} from './queryAndBuildAssetBaseOSSData';
+import {ApolloClient} from '../apollo-client';
+
+export const queryAndBuildAssetBaseData = async (
+  keys: string[],
+  client: ApolloClient<any>,
+): Promise<Record<string, LiveDataForNode>> => {
+  return queryAndBuildAssetBaseOSSData(keys, client);
+};
