+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/LiveDataForNode.oss.tsx
@@ -1,0 +1,6 @@
+import {
+  OSSLiveDataForNode as LiveDataForNode,
+  OSSLiveDataForNodeWithStaleData as LiveDataForNodeWithStaleData,
+} from './OSSLiveDataForNode';
+
+export type {LiveDataForNode, LiveDataForNodeWithStaleData};
