+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/MissingLiveData.oss.tsx
@@ -1,0 +1,5 @@
+import {LiveDataForNodeWithStaleData} from 'shared/asset-graph/LiveDataForNode.oss';
+
+import {MISSING_LIVE_DATA_OSS} from './MissingLiveDataOSS';
+
+export const MISSING_LIVE_DATA: LiveDataForNodeWithStaleData = MISSING_LIVE_DATA_OSS;
