+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/MissingLiveDataOSS.tsx
@@ -1,0 +1,17 @@
+import {OSSLiveDataForNodeWithStaleData} from './OSSLiveDataForNode';
+
+export const MISSING_LIVE_DATA_OSS: OSSLiveDataForNodeWithStaleData = {
+  unstartedRunIds: [],
+  inProgressRunIds: [],
+  runWhichFailedToMaterialize: null,
+  freshnessInfo: null,
+  lastMaterialization: null,
+  lastMaterializationRunStatus: null,
+  lastObservation: null,
+  partitionStats: null,
+  staleStatus: null,
+  staleCauses: [],
+  assetChecks: [],
+  opNames: [],
+  stepKey: '',
+};
