+++ js_modules/dagster-ui/packages/ui-core/src/assets/FailedRunSinceMaterializationBanner.tsx
@@ -6,7 +6,7 @@ import {
 } from '@dagster-io/ui-components/src/components/types';
 import {Link} from 'react-router-dom';
 
-import {AssetLatestInfoRunFragment} from '../asset-data/types/AssetBaseDataProvider.types';
+import {AssetLatestInfoRunFragment} from '../asset-data/types/AssetBaseDataQueries.types';
 import {RunStatus} from '../graphql/types';
 import {titleForRun} from '../runs/RunUtils';
 import {useStepLogs} from '../runs/StepLogsDialog';
