+++ js_modules/dagster-ui/packages/ui-core/src/assets/asset-checks/AssetChecksStatusSummary.tsx
@@ -10,11 +10,11 @@ import {
 } from '@dagster-io/ui-components';
 import countBy from 'lodash/countBy';
 import * as React from 'react';
+import {LiveDataForNode} from 'shared/asset-graph/LiveDataForNode.oss';
 
 import {CheckStatusRow} from './AssetCheckStatusTag';
 import {assertUnreachable} from '../../app/Util';
-import {AssetCheckLiveFragment} from '../../asset-data/types/AssetBaseDataProvider.types';
-import {LiveDataForNode} from '../../asset-graph/Utils';
+import {AssetCheckLiveFragment} from '../../asset-data/types/AssetBaseDataQueries.types';
 import {
   AssetCheckExecutionResolvedStatus,
   AssetCheckSeverity,
