+++ js_modules/dagster-ui/packages/ui-core/src/graphql/types.ts
@@ -980,6 +980,9 @@ export enum DagsterEventType {
   RUN_START = 'RUN_START',
   RUN_STARTING = 'RUN_STARTING',
   RUN_SUCCESS = 'RUN_SUCCESS',
+  SLA_PASSING = 'SLA_PASSING',
+  SLA_UNKNOWN = 'SLA_UNKNOWN',
+  SLA_VIOLATING = 'SLA_VIOLATING',
   STEP_EXPECTATION_RESULT = 'STEP_EXPECTATION_RESULT',
   STEP_FAILURE = 'STEP_FAILURE',
   STEP_INPUT = 'STEP_INPUT',
